requirements.txt
streamlit>=1.29.0
selenium>=4.15.2
beautifulsoup4>=4.12.2
google-generativeai>=0.3.0
python-dotenv>=1.0.0
webdriver-manager>=4.0.1
pandas>=2.1.3
